@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-#  Copyright (c) 2014, Pfadibewegung Schweiz. This file is part of
+#  Copyright (c) 2014-2017, Pfadibewegung Schweiz. This file is part of
 #  hitobito_youth and licensed under the Affero General Public License version 3
 #  or later. See the COPYING file at the top-level directory or at
 #  https://github.com/hitobito/hitobito_youth.
@@ -12,7 +12,6 @@ module Youth::FilterNavigation::People
     alias_method_chain :path, :education
     alias_method_chain :new_group_people_filter_path, :education
     alias_method_chain :delete_group_people_filter_path, :education
-    alias_method_chain :qualification_group_people_filter_path, :education
   end
 
   private
@@ -27,9 +26,11 @@ module Youth::FilterNavigation::People
 
   def new_group_people_filter_path_with_education
     template.new_group_people_filter_path(
-      group.id,
-      education: education?,
-      people_filter: { role_type_ids: role_type_ids })
+        group.id,
+        education: education?,
+        range: filter.range,
+        filters: filter.chain.to_params
+    )
   end
 
   def delete_group_people_filter_path_with_education(filter)
